Added a Get action, and some other fixes

+++ src/main/java/org/elasticsearch/service/DynamicPathService.java
@@ -1,14 +1,17 @@
 package org.elasticsearch.service;
 
 import java.io.IOException;
+import java.nio.file.FileSystemException;
+import java.nio.file.FileSystemNotFoundException;
 import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.security.AccessController;
-import java.security.PrivilegedAction;
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.apache.logging.log4j.LogManager;
@@ -24,15 +27,16 @@ public class DynamicPathService {
     configDirectory = environment.configFile().toAbsolutePath();
   }
 
-  public void appendLine(String path, String line) throws IOException {
+  public void appendLine(String path, String line) throws FileSystemException {
     // Append line to path
     Path absPath = getAbsolutePath(path);
     String lineWithNewLine = line.concat("\r\n");
 
-    LOGGER.debug("Adding line {} to path {}", line, path);
+    LOGGER.debug("Adding line {} to path {}", line, absPath);
 
     doPrivileged(() -> {
       if (!Files.exists(absPath)) {
+        Files.createDirectories(absPath.getParent());
         Files.write(absPath, lineWithNewLine.getBytes(), StandardOpenOption.CREATE);
       } else {
         Files.write(absPath, lineWithNewLine.getBytes(), StandardOpenOption.APPEND);
@@ -42,16 +46,68 @@ public class DynamicPathService {
     });
   }
 
-  public void removeLine(String path, String line) throws IOException {
+  public void removeLine(String path, String line) throws FileSystemException {
     Path absPath = getAbsolutePath(path);
 
-    LOGGER.debug("Removing line {} from path {}", line, path);
+    LOGGER.debug("Removing line {} from path {}", line, absPath);
 
     doPrivileged(() -> {
-      List<String> filtered = Files.lines(absPath).filter(l -> !linesAreEqual(l, line)).collect(Collectors.toList());
-      Files.write(absPath, filtered);
+      if (Files.exists(absPath)) {
+        List<String> filtered = Files.lines(absPath).filter(l -> !linesAreEqual(l, line)).collect(Collectors.toList());
+        Files.write(absPath, filtered);
+      } else {
+        LOGGER.warn("Tried to remove line from {}, but file does not exist", absPath);
+      }
+
       return null;
     });
+  }
+
+  public void createFile(String path, String line) throws FileSystemException {
+    Path absPath = getAbsolutePath(path);
+    String lineWithNewLine = line.concat("\r\n");
+
+    doPrivileged(() -> {
+      if (!Files.exists(absPath)) {
+        LOGGER.debug("Creating file {}", absPath);
+        Files.createDirectories(absPath.getParent());
+        Files.write(absPath, lineWithNewLine.getBytes(), StandardOpenOption.CREATE);
+      } else {
+        LOGGER.debug("Not creating file {}, it already exists", absPath);
+      }
+
+      return null;
+    });
+  }
+
+  public List<String> getLines(String path) throws FileSystemException {
+    Path absPath = getAbsolutePath(path);
+
+    LOGGER.debug("Getting lines from path {}", absPath);
+
+    try {
+      return AccessController.doPrivileged((PrivilegedExceptionAction<List<String>>) () -> {
+        if (Files.exists(absPath)) {
+          return Files.lines(absPath)
+              .filter(l -> !lineIsComment(l))
+              .collect(Collectors.toList());
+        } else {
+          LOGGER.warn("Cannot get line for path {}: file does not exist", absPath);
+          throw new FileSystemNotFoundException();
+        }
+      });
+    } catch (PrivilegedActionException e) {
+      if (e.getException() instanceof FileSystemException) {
+        throw (FileSystemException) e.getException();
+      } else {
+        e.printStackTrace();
+        return new ArrayList<>();
+      }
+    }
+  }
+
+  private boolean lineIsComment(String line) {
+    return line.startsWith("#");
   }
 
   private boolean linesAreEqual(String lhs, String rhs) {
@@ -68,15 +124,13 @@ public class DynamicPathService {
     return Paths.get(configDirectory.toString(), path);
   }
 
-  private void doPrivileged(PrivilegedExceptionAction<Void> action) throws IOException {
+  private void doPrivileged(PrivilegedExceptionAction<Void> action) throws FileSystemException {
     try {
       AccessController.doPrivileged(action);
     } catch (PrivilegedActionException e) {
-      if (e.getException() instanceof IOException) {
-        throw (IOException) e.getException();
+      if (e.getException() instanceof FileSystemException) {
+        throw (FileSystemException) e.getException();
       }
-
-      e.printStackTrace();
     }
   }
